pyproject.toml
[tool.poetry]
name = "nonebot-plugin-lolinfo"
version = "0.2.5"
description = "以图片形式将LOL的各种数据发送到QQ中"
authors = ["Shadow403"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10,<4.0"
httpx = ">=0.27.0"
nonebot2 = ">=2.3.0"
pydantic = ">=2.8.0"
nonebot-adapter-onebot = ">=2.4.0"
nonebot-plugin-htmlrender = ">=0.3.0"
nonebot-plugin-send-anything-anywhere = "^0.7.1"
pillow = "^11.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E402"]